refactor(api): unify v3 routing and split server wiring
