+++ internal/api/server_paths.go
@@ -1,0 +1,70 @@
+// Licensed under the PolyForm Noncommercial License 1.0.0
+// Since v2.0.0, this software is restricted to non-commercial use only.
+
+package api
+
+import (
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+)
+
+// dataFilePath resolves a relative path inside the configured data directory while
+// protecting against path traversal and symlink escapes. The returned path points to
+// the real location on disk and is safe to open.
+func (s *Server) dataFilePath(rel string) (string, error) {
+	clean := filepath.Clean(rel)
+	if filepath.IsAbs(clean) {
+		return "", fmt.Errorf("data file path must be relative: %s", rel)
+	}
+	if strings.Contains(clean, "..") {
+		return "", fmt.Errorf("data file path contains traversal: %s", rel)
+	}
+
+	s.mu.RLock()
+	dataDir := s.cfg.DataDir
+	s.mu.RUnlock()
+
+	root, err := filepath.Abs(dataDir)
+	if err != nil {
+		return "", fmt.Errorf("resolve data directory: %w", err)
+	}
+
+	full := filepath.Join(root, clean)
+	resolvedRoot, err := filepath.EvalSymlinks(root)
+	if err != nil {
+		resolvedRoot = root
+	}
+
+	resolved := full
+	if info, statErr := os.Stat(full); statErr == nil {
+		if info.IsDir() {
+			return "", fmt.Errorf("data file path points to directory: %s", rel)
+		}
+		if resolvedPath, evalErr := filepath.EvalSymlinks(full); evalErr == nil {
+			resolved = resolvedPath
+		}
+	} else if !errors.Is(statErr, os.ErrNotExist) {
+		return "", fmt.Errorf("stat data file: %w", statErr)
+	} else {
+		// File might be generated later; still ensure parent directories stay within root.
+		dir := filepath.Dir(full)
+		if _, dirErr := os.Stat(dir); dirErr == nil {
+			if realDir, evalErr := filepath.EvalSymlinks(dir); evalErr == nil {
+				resolved = filepath.Join(realDir, filepath.Base(full))
+			}
+		}
+	}
+
+	relToRoot, err := filepath.Rel(resolvedRoot, resolved)
+	if err != nil {
+		return "", fmt.Errorf("resolve relative path: %w", err)
+	}
+	if strings.HasPrefix(relToRoot, "..") || filepath.IsAbs(relToRoot) {
+		return "", fmt.Errorf("data file escapes data directory: %s", rel)
+	}
+
+	return resolved, nil
+}
