+++ internal/api/server_refresh.go
@@ -1,0 +1,156 @@
+// Licensed under the PolyForm Noncommercial License 1.0.0
+// Since v2.0.0, this software is restricted to non-commercial use only.
+
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/ManuGH/xg2g/internal/jobs"
+	"github.com/ManuGH/xg2g/internal/log"
+	"github.com/ManuGH/xg2g/internal/resilience"
+)
+
+// HandleRefreshInternal exposes the refresh handler for versioned APIs
+// This allows different API versions to wrap the core refresh logic
+func (s *Server) HandleRefreshInternal(w http.ResponseWriter, r *http.Request) {
+	s.handleRefresh(w, r)
+}
+
+func (s *Server) handleRefresh(w http.ResponseWriter, r *http.Request) {
+	logger := log.WithComponentFromContext(r.Context(), "api")
+	actor := r.RemoteAddr
+
+	// Try to acquire the refresh flag atomically; fail fast if already running
+	if !s.refreshing.CompareAndSwap(false, true) {
+		logger.Warn().
+			Str("event", "refresh.conflict").
+			Str("method", r.Method).
+			Msg("refresh already in progress")
+
+		w.Header().Set("Content-Type", "application/json")
+		w.Header().Set("Retry-After", "30") // suggest retry after 30s
+		w.WriteHeader(http.StatusConflict)
+		_ = json.NewEncoder(w).Encode(map[string]string{
+			"error":  "conflict",
+			"detail": "A refresh operation is already in progress",
+		})
+		return
+	}
+	defer s.refreshing.Store(false)
+
+	// Capture snapshot once to prevent config drift within this operation.
+	s.mu.RLock()
+	snap := s.snap
+	s.mu.RUnlock()
+
+	// Audit log: refresh started
+	bouquets := strings.Split(snap.App.Bouquet, ",")
+	if s.auditLogger != nil {
+		s.auditLogger.RefreshStart(actor, bouquets)
+	}
+
+	// Create independent context for background job
+	// Use Background() instead of request context to prevent premature cancellation
+	jobCtx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	defer cancel()
+
+	// Optional: Monitor client disconnect for logging
+	clientDisconnected := make(chan struct{})
+	go func() {
+		<-r.Context().Done()
+		if r.Context().Err() == context.Canceled {
+			logger.Info().Msg("client disconnected during refresh (job continues)")
+			close(clientDisconnected)
+		}
+	}()
+
+	start := time.Now()
+	var st *jobs.Status
+	// Run the refresh via circuit breaker; it will mark failures and handle panics
+	err := s.cb.Execute(func() error {
+		var err error
+		st, err = s.refreshFn(jobCtx, snap)
+		return err
+	})
+	duration := time.Since(start)
+
+	w.Header().Set("Content-Type", "application/json")
+	if err != nil {
+		// Audit log: refresh error
+		if s.auditLogger != nil {
+			s.auditLogger.RefreshError(actor, err.Error())
+		}
+
+		// Distinguish open circuit (fast-fail) from internal error
+		if errors.Is(err, resilience.ErrCircuitOpen) {
+			logger.Warn().
+				Str("event", "refresh.circuit_open").
+				Int64("duration_ms", duration.Milliseconds()).
+				Msg("circuit breaker open for refresh; rejecting request")
+			w.Header().Set("Content-Type", "application/json")
+			w.WriteHeader(http.StatusServiceUnavailable)
+			_ = json.NewEncoder(w).Encode(map[string]string{
+				"error":  "unavailable",
+				"detail": "Refresh temporarily disabled due to repeated failures",
+			})
+			return
+		}
+		s.mu.Lock()
+		s.status.Error = "refresh operation failed" // Security: don't expose internal error details
+		s.status.Channels = 0                       // NEW: reset channel count on error
+		s.mu.Unlock()
+
+		logger.Error().
+			Err(err).
+			Str("event", "refresh.failed").
+			Str("method", r.Method).
+			Int64("duration_ms", duration.Milliseconds()).
+			Str("status", "error").
+			Msg("refresh failed")
+		// Security: Never expose internal error details to client
+		http.Error(w, "Internal server error", http.StatusInternalServerError)
+		return
+	}
+
+	// Audit log: refresh completed successfully
+	if s.auditLogger != nil {
+		s.auditLogger.RefreshComplete(actor, st.Channels, st.Bouquets, duration.Milliseconds())
+	}
+
+	recordRefreshMetrics(duration, st.Channels)
+
+	select {
+	case <-clientDisconnected:
+		logger.Info().
+			Str("event", "refresh.success").
+			Str("method", r.Method).
+			Int("channels", st.Channels).
+			Int64("duration_ms", duration.Milliseconds()).
+			Str("status", "success").
+			Msg("refresh completed despite client disconnect")
+	default:
+		logger.Info().
+			Str("event", "refresh.success").
+			Str("method", r.Method).
+			Int("channels", st.Channels).
+			Int64("duration_ms", duration.Milliseconds()).
+			Str("status", "success").
+			Msg("refresh completed successfully")
+	}
+
+	s.mu.Lock()
+	s.status = *st
+	s.mu.Unlock()
+
+	if err := json.NewEncoder(w).Encode(st); err != nil {
+		logger.Error().Err(err).Str("event", "refresh.encode_error").Msg("failed to encode refresh response")
+		http.Error(w, "Internal server error", http.StatusInternalServerError)
+		return
+	}
+}
